Ensure observation reviews get created for the observer if there's a taxon.
This was working based on identifications until I recently made it so idents
don't add reviews if they've been instructed to ignore updating the observation,
which makes sense to me and lets us skip one more of the kajillion callbacks
that get triggered when you create an identitfication if we so desire. This
restores the old behavior while keeping that option to skip.